refactor: split marshal/variant_reconstruct.go — standalone helpers into variant_reconstruct_helpers.go
